requirements.txt
streamlit
Pillow
requests
openai
